Rename serviceFacade files to remove stutter from imports
