+++ src/components/data/DataNavigation.js
@@ -38,7 +38,7 @@ import ArrowRightIcon from "@material-ui/icons/ArrowRight";
 import FolderIcon from "@material-ui/icons/Folder";
 import ids from "./ids";
 import constants from "../../constants";
-import { getFilesystemRoots } from "../../serviceFacade/filesystemServiceFacade";
+import { getFilesystemRoots } from "../../serviceFacades/filesystem";
 import { useUserProfile } from "../../contexts/userProfile";
 import { useQuery } from "react-query";
