+++ src/serviceFacades/sharing.js
